+++ R/adlib.R
@@ -111,7 +111,7 @@ adlib_build_query <- function(ad_reached_countries,
                               publisher_platform = "FACEBOOK",
                               search_page_ids = NULL,
                               search_terms = NULL,
-                              limit = 5000,
+                              limit = 1000,
                               fields = "ad_data") {
   ad_active_status <- match.arg(ad_active_status)
   ad_type <- match.arg(ad_type)
@@ -280,18 +280,7 @@ format_array <- function(items) {
 }
 
 
-#' Convert Ad Library response to tibble
-#'
-#' @param response a response form adlib_get
-#'
-#' @return tibble
 #' @export
-#'
 #' @importFrom tibble as_tibble
-#' @importFrom purrr map
-#' @importFrom dplyr bind_rows
-adlib_data_frame <- function(response) {
-  adlib_data <- content(response)[["data"]]
-  map(adlib_data, as_tibble) %>%
-    bind_rows()
-}
+tibble::as_tibble
+