+++ NEWS.md
@@ -1,3 +1,11 @@
+# Radlibrary 0.2.0
+* CI is set up and passes
+* API Requests set the user agent to Radlibrary
+* Access tokens are censored in URLs by default
+* Spend and Impressions columns have numeric upper and lower bounds
+* Documentation
+* Bug fixes
+
 # Radlibrary 0.1.0
 * Secrets are stored securely in keychain
 * Requests return data_objects that can be converted to tibbles